Move clap App construction into separate module
Also join two parts of the construction into a single function
